shift bootstrap details into own file & update build system
