+++ lua/crowlib.lua
@@ -1,3 +1,17 @@
+-- temp placing print redef here to complete bootstrap elimination
+print = function(...)
+    local args = {...}
+    local arg_len = #args
+    local printResult = args[1]
+    if arg_len > 1 then
+        for i=2,arg_len do
+            printResult = printResult .. '\t' .. tostring(args[i])
+        end
+    end
+    print_serial(tostring(printResult))
+end
+
+
 --- Crow standard library
 
 local C = {}
